02_esdl_updater/requirements.txt
pyecore
pyESDL
requests
pymysql